+++ DotfilesLibrary/DotfilesLibrary.py
@@ -1,14 +1,17 @@
-from .Link import Link
-from . import Misc
+from DotfilesLibrary.Install import Install
+from DotfilesLibrary.Link import Link
+from DotfilesLibrary import Interactive, Emit
 
 class DotfilesLibrary:
     def __init__(self, *args, **kwargs):
         self._link = Link(*args, **kwargs)
+        self._install = Install()
 
         # TODO: Use reflection to get functions from all modules.
         self._keywords = [
-            Misc.interactive,
-            Misc.emit,
+            Interactive.interactive,
+            Emit.emit,
+            self._install.install,
             self._link.set_cwd,
             self._link.set_target,
             self._link.add_ignore,